Corrección del bug
Se incluye en el nuevo fichero "commons.php" la funcionalidad por
completo del formulario, incluyendo la nueva que permite controlar
mediante token el envío.
